перенес работу с консолью с отдельный класс
