+++ Bank/TypeEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace Bank;
+
+enum TypeEnum
+{
+    VIP = 1, //особые условие
+    Normal = 2, //обычные условия
+    Corp = 3, //корпоративные условия(зарплатная карта)
+    Comp = 4, //счет компании
+}
