--- HW_2/Account/TypeEnum.cs
@@ -1,10 +1,0 @@
-﻿namespace OOP_Basics.Account
-{
-    enum TypeEnum
-    {
-        VIP = 1, //особые условие
-        Normal = 2, //обычные условия
-        Corp = 3, //корпоративные условия(зарплатная карта)
-        Comp = 4, //счет компании
-    }
-}
