+++ HW_2/Program.cs
@@ -1,4 +1,4 @@
-﻿using OOP_Basics.Account;
+﻿using Bank;
 
 
 static string StringReverse(string source)
