--- HW_2/task3.txt
@@ -1,12 +1,0 @@
-﻿1. В класс банковский счет, созданный в упражнениях, добавить метод, 
-который переводит деньги с одного счета на другой. У метода два параметра:
-ссылка на объект класса банковский счет откуда снимаются деньги, второй параметр – сумма.
-
-2. Реализовать метод, который в качестве входного параметра принимает строку string,
-возвращает строку типа string, буквы в которой идут в обратном порядке. Протестировать метод.
-
-3. * Работа со строками. Дан текстовый файл, содержащий ФИО и e-mail адрес. 
-Разделителем между ФИО и адресом электронной почты является символ &: 
-Кучма Андрей Витальевич & Kuchma@mail.ru Мизинцев Павел Николаевич & Pasha@mail.ru Сформировать новый файл,
-содержащий список адресов электронной почты. Предусмотреть метод, выделяющий из строки адрес почты.
-Методу в качестве параметра передается символьная строка s, e-mail возвращается в той же строке s: public void SearchMail (ref string s).
