+++ OOP_Basics.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.2.32519.379
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OOP_Basics", "HW_2\OOP_Basics.csproj", "{FF634E79-1E7D-4F34-898D-FDC15167E928}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bank", "Bank\Bank.csproj", "{17554ABA-9811-436E-9583-3B1E2B3EFA7A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{FF634E79-1E7D-4F34-898D-FDC15167E928}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF634E79-1E7D-4F34-898D-FDC15167E928}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FF634E79-1E7D-4F34-898D-FDC15167E928}.Release|Any CPU.Build.0 = Release|Any CPU
+		{17554ABA-9811-436E-9583-3B1E2B3EFA7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{17554ABA-9811-436E-9583-3B1E2B3EFA7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{17554ABA-9811-436E-9583-3B1E2B3EFA7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{17554ABA-9811-436E-9583-3B1E2B3EFA7A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
